% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{add_time_column}
\alias{add_time_column}
\title{Helper function to add time column if missing}
\arguments{
\item{data}{A data.frame.}

\item{id_col}{Character string. The name of the ID column to group by when
creating the time index. Defaults to "id".}

\item{time_col}{Character string. The name of the time column to create.
Defaults to "time".}
}
\value{
The data.frame with the time column added if it was missing.
}
\description{
This internal helper function adds a sequential time column to a data frame
if it's missing, grouping by an ID column if specified. This is useful for
plotting functions that require a time index.
}
\keyword{internal}
