% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{calculate_native_distance}
\alias{calculate_native_distance}
\title{Calculate all available native distances}
\usage{
calculate_native_distance(x, y, method, ...)
}
\arguments{
\item{x}{First vector}

\item{y}{Second vector}

\item{method}{Distance method name}

\item{...}{Additional parameters}
}
\value{
Distance value
}
\description{
Calculate all available native distances
}
