% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{calculate_state_statistics}
\alias{calculate_state_statistics}
\title{Calculate comprehensive statistics for states}
\usage{
calculate_state_statistics(states, values, data, value_column)
}
\arguments{
\item{states}{Vector of state assignments}

\item{values}{Original time series values}

\item{data}{Original data frame}

\item{value_column}{Name of value column}
}
\value{
List containing state statistics and summary
}
\description{
Calculate comprehensive statistics for states
}
