% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator.R
\name{calculate_unified_distance}
\alias{calculate_unified_distance}
\title{Unified distance calculation}
\usage{
calculate_unified_distance(x, y, method, ...)
}
\arguments{
\item{x}{First vector}

\item{y}{Second vector}

\item{method}{Distance method}

\item{...}{Additional parameters}
}
\value{
Distance value
}
\description{
Unified distance calculation
}
