% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{canberra_distance}
\alias{canberra_distance}
\title{Calculate Canberra distance between two vectors}
\usage{
canberra_distance(x, y)
}
\arguments{
\item{x}{First vector}

\item{y}{Second vector}
}
\value{
Canberra distance
}
\description{
Calculate Canberra distance between two vectors
}
