% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{chebyshev_distance}
\alias{chebyshev_distance}
\title{Calculate Chebyshev (maximum) distance between two vectors}
\usage{
chebyshev_distance(x, y)
}
\arguments{
\item{x}{First vector}

\item{y}{Second vector}
}
\value{
Chebyshev distance
}
\description{
Calculate Chebyshev (maximum) distance between two vectors
}
