% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_required_packages}
\alias{check_required_packages}
\title{Check Required Packages}
\usage{
check_required_packages(pkgs, fn_name)
}
\arguments{
\item{pkgs}{Character vector of package names to check.}

\item{fn_name}{Character string. The name of the function calling this check,
used for more informative error messages.}
}
\description{
Checks if specified packages are installed. If not, it stops execution
with an informative error message.
}
\keyword{internal}
