% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{correlation_distance}
\alias{correlation_distance}
\title{Calculate correlation-based distance}
\usage{
correlation_distance(x, y, method = "pearson")
}
\arguments{
\item{x}{First vector}

\item{y}{Second vector}

\item{method}{Correlation method ("pearson", "spearman", "kendall")}
}
\value{
Correlation distance (1 - |correlation|)
}
\description{
Calculate correlation-based distance
}
