% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{cosine_distance}
\alias{cosine_distance}
\title{Calculate Cosine distance between two vectors}
\usage{
cosine_distance(x, y)
}
\arguments{
\item{x}{First vector}

\item{y}{Second vector}
}
\value{
Cosine distance (1 - cosine similarity)
}
\description{
Calculate Cosine distance between two vectors
}
