% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_formatting.R
\name{create_windowed_network_output}
\alias{create_windowed_network_output}
\title{Create windowed analysis network output}
\usage{
create_windowed_network_output(segments, distances, method_name, output_format)
}
\arguments{
\item{segments}{List of segments with metadata}

\item{distances}{Vector of calculated distances}

\item{method_name}{Method used for calculation}

\item{output_format}{Format of the output, either "matrix" or "edges".}
}
\description{
Create windowed analysis network output
}
