% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{dist_ccf}
\alias{dist_ccf}
\title{Calculate cross-correlation function between time series}
\usage{
dist_ccf(series1, series2, type = "max_absolute", max_lag = NULL)
}
\arguments{
\item{series1}{First time series}

\item{series2}{Second time series}

\item{type}{Type of CCF to return ("max_absolute", "max_positive", "min_negative", "full_sequence")}

\item{max_lag}{Maximum lag to consider}
}
\value{
List with CCF value and lag
}
\description{
Calculate cross-correlation function between time series
}
