% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{dist_dtw}
\alias{dist_dtw}
\title{Calculate Dynamic Time Warping distance}
\usage{
dist_dtw(series1, series2, ...)
}
\arguments{
\item{series1}{First time series}

\item{series2}{Second time series}

\item{...}{Additional arguments for DTW}
}
\value{
List with distance
}
\description{
Calculate Dynamic Time Warping distance
}
