% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{dist_mic}
\alias{dist_mic}
\title{Calculate Maximal Information Coefficient}
\usage{
dist_mic(series1, series2, ...)
}
\arguments{
\item{series1}{First time series}

\item{series2}{Second time series}

\item{...}{Additional arguments for MIC}
}
\value{
List with MIC value
}
\description{
Calculate Maximal Information Coefficient
}
