% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{dist_voi}
\alias{dist_voi}
\title{Calculate Variation of Information}
\usage{
dist_voi(
  series1,
  series2,
  n_bins = 10,
  discretize_method = "quantile",
  already_discrete = FALSE
)
}
\arguments{
\item{series1}{First time series}

\item{series2}{Second time series}

\item{n_bins}{Number of bins for discretization}

\item{discretize_method}{Discretization method}

\item{already_discrete}{Whether series are already discrete}
}
\value{
List with VI value
}
\description{
Calculate Variation of Information
}
