% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{.jitter_spike_train}
\alias{.jitter_spike_train}
\title{Jitter spike train for significance testing}
\usage{
.jitter_spike_train(spike_train, jitter_val, series_length_for_jitter = NULL)
}
\arguments{
\item{spike_train}{Numeric vector of spike times}

\item{jitter_val}{Numeric. The range for jittering.}

\item{series_length_for_jitter}{Numeric. The maximum possible time point in the series.}
}
\value{
Jittered spike train
}
\description{
Jitter spike train for significance testing
}
\keyword{internal}
