% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{dtw_distance}
\alias{dtw_distance}
\title{Native DTW implementation (simplified)}
\usage{
dtw_distance(x, y)
}
\arguments{
\item{x}{First vector}

\item{y}{Second vector}
}
\value{
DTW distance
}
\description{
Native DTW implementation (simplified)
}
