% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Calculate Euclidean distance between two vectors}
\usage{
euclidean_distance(x, y)
}
\arguments{
\item{x}{First vector}

\item{y}{Second vector}
}
\value{
Euclidean distance
}
\description{
Calculate Euclidean distance between two vectors
}
