% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator.R
\name{extract_ts_from_dataframe}
\alias{extract_ts_from_dataframe}
\title{Extract time series and ID information from data frame}
\usage{
extract_ts_from_dataframe(ts_data, ts_col_name, id_col_name)
}
\arguments{
\item{ts_data}{The data frame containing the time series and ID data.}

\item{ts_col_name}{The name of the column containing the time series data.}

\item{id_col_name}{The name of the column containing the ID data.}
}
\value{
A list containing the time series vector and a list of segments with their IDs, start and end indices.
}
\description{
Extract time series and ID information from data frame
}
