% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator.R
\name{extract_ts_vector}
\alias{extract_ts_vector}
\title{Extract time series vector from various input formats}
\usage{
extract_ts_vector(ts_data, ts_col_name = NULL)
}
\arguments{
\item{ts_data}{Time series data, can be a vector or a data frame.}

\item{ts_col_name}{The name of the column containing the time series data if ts_data is a data frame.}
}
\value{
A numeric vector of the time series.
}
\description{
Extract time series vector from various input formats
}
