% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator.R
\name{get_all_methods}
\alias{get_all_methods}
\title{Get all available distance methods}
\usage{
get_all_methods()
}
\value{
Vector of all available method names
}
\description{
Get all available distance methods
}
