% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{get_native_methods}
\alias{get_native_methods}
\title{Get list of all available native distance methods}
\usage{
get_native_methods()
}
\value{
Vector of method names
}
\description{
Get list of all available native distance methods
}
