% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_utils.R
\name{get_scaling_summary}
\alias{get_scaling_summary}
\title{Get Scaling Summary}
\usage{
get_scaling_summary(scaled_object)
}
\arguments{
\item{scaled_object}{Scaled data object (from apply_scaling or scale_dataframe)}
}
\value{
Summary of scaling information
}
\description{
Provides a summary of scaling transformations applied to data.
}
\examples{
data <- rnorm(100, mean = 50, sd = 10)
scaled_result <- apply_scaling(data, method = "standardize")
get_scaling_summary(scaled_result)

}
