% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{manhattan_distance}
\alias{manhattan_distance}
\title{Calculate Manhattan (L1) distance between two vectors}
\usage{
manhattan_distance(x, y)
}
\arguments{
\item{x}{First vector}

\item{y}{Second vector}
}
\value{
Manhattan distance
}
\description{
Calculate Manhattan (L1) distance between two vectors
}
