% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{minkowski_distance}
\alias{minkowski_distance}
\title{Calculate Minkowski distance between two vectors}
\usage{
minkowski_distance(x, y, p = 2)
}
\arguments{
\item{x}{First vector}

\item{y}{Second vector}

\item{p}{Power parameter (p=1: Manhattan, p=2: Euclidean)}
}
\value{
Minkowski distance
}
\description{
Calculate Minkowski distance between two vectors
}
