% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator.R
\name{perform_id_partition_analysis}
\alias{perform_id_partition_analysis}
\title{Perform ID-based partition analysis}
\usage{
perform_id_partition_analysis(ts_info, method, output_format, pairwise, ...)
}
\arguments{
\item{ts_info}{A list containing the time series and segment information.}

\item{method}{The distance calculation method.}

\item{output_format}{The format of the output, either "matrix" or "edges".}

\item{pairwise}{If TRUE, all pairwise distances are calculated; if FALSE, only consecutive segments are compared.}

\item{...}{Additional arguments passed to the distance functions.}
}
\value{
A list containing the results of the ID-based partition analysis.
}
\description{
Perform ID-based partition analysis
}
