% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurst_analysis.R
\name{print.hurst_result}
\alias{print.hurst_result}
\title{Print method for hurst_result objects}
\usage{
\method{print}{hurst_result}(x, ...)
}
\arguments{
\item{x}{A hurst_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for hurst_result objects
}
