% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resilience_states.R
\name{print.resilience_diagnosis}
\alias{print.resilience_diagnosis}
\title{Print method for resilience_diagnosis objects}
\usage{
\method{print}{resilience_diagnosis}(x, ...)
}
\arguments{
\item{x}{A resilience_diagnosis object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for resilience_diagnosis objects
}
