% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_utils.R
\name{print.scaled_data}
\alias{print.scaled_data}
\title{Print method for scaled_data objects}
\usage{
\method{print}{scaled_data}(x, ...)
}
\arguments{
\item{x}{A scaled_data object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for scaled_data objects
}
