% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_utils.R
\name{print.scaled_dataframe}
\alias{print.scaled_dataframe}
\title{Print method for scaled_dataframe objects}
\usage{
\method{print}{scaled_dataframe}(x, ...)
}
\arguments{
\item{x}{A scaled_dataframe object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for scaled_dataframe objects
}
