% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{reverse_transform}
\alias{reverse_transform}
\title{Reverse transform time series data}
\usage{
reverse_transform(transformed, method, parameters)
}
\arguments{
\item{transformed}{Transformed series}

\item{method}{Original transformation method}

\item{parameters}{Parameters used in the original transformation}
}
\value{
Original series
}
\description{
Reverse transform time series data
}
