% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{transform_series}
\alias{transform_series}
\title{Transform time series data}
\usage{
transform_series(series, method = "none")
}
\arguments{
\item{series}{Numeric vector of time series values}

\item{method}{Transformation method ("none", "log", "sqrt", "scale", "center")}
}
\value{
List with transformed series and transformation info
}
\description{
Transform time series data
}
