% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_distance.R
\name{ts_distance_methods}
\alias{ts_distance_methods}
\title{Get Available Distance Methods}
\usage{
ts_distance_methods()
}
\value{
Character vector of method names
}
\description{
Returns list of all available distance calculation methods
}
