% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_utils.R
\name{ts_info}
\alias{ts_info}
\title{Package Information}
\usage{
ts_info()
}
\value{
A list of categorized function names, printed to the console.
}
\description{
Displays an overview of the main functions and capabilities provided by the
`tsnetworks` package. This serves as a quick reference for users.
}
