% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_network.R
\name{ts_network_recommend}
\alias{ts_network_recommend}
\title{Recommend Network Method for Distance Scale}
\usage{
ts_network_recommend(distance_matrix)
}
\arguments{
\item{distance_matrix}{Distance matrix}
}
\value{
List with recommended method and parameters
}
\description{
Suggests appropriate network method based on distance distribution
}
