% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_dataframe_input}
\alias{validate_dataframe_input}
\title{Validate data frame input for time series}
\usage{
validate_dataframe_input(data, series_col_name = NULL, id_col_name = NULL)
}
\arguments{
\item{data}{Data frame}

\item{series_col_name}{Column name for time series}

\item{id_col_name}{Column name for identifiers}
}
\description{
Validate data frame input for time series
}
