% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_method}
\alias{validate_method}
\title{Validate method parameters}
\usage{
validate_method(method_name, available_methods = NULL)
}
\arguments{
\item{method_name}{Distance method name}

\item{available_methods}{Vector of available methods}
}
\description{
Validate method parameters
}
