% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_time_series}
\alias{validate_time_series}
\title{Validate time series input}
\usage{
validate_time_series(series, series_name = "series", allow_empty = FALSE)
}
\arguments{
\item{series}{Time series data}

\item{series_name}{Name for error messages}

\item{allow_empty}{Whether to allow empty series}
}
\value{
Validated series
}
\description{
Validate time series input
}
