% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_window_params}
\alias{validate_window_params}
\title{Validate window parameters}
\usage{
validate_window_params(window_size, step_size = NULL, series_length = NULL)
}
\arguments{
\item{window_size}{Window size}

\item{step_size}{Step size}

\item{series_length}{Length of the series}
}
\description{
Validate window parameters
}
