% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var9}
\alias{var9}
\title{VAR(9) Time Series Dataset}
\format{
A data frame with time series observations and multiple variables
}
\source{
Simulated data from a VAR(9) model for testing purposes
}
\usage{
var9
}
\description{
A multivariate time series dataset generated from a Vector Autoregressive model
of order 9 (VAR(9)). This dataset is used for demonstrating multivariate time
series network analysis and testing various distance measures in the tsnetworks package.
}
\examples{
data(var9)
head(var9)
# Basic summary of the dataset
summary(var9)
}
\keyword{datasets}
